% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{addQuotes}
\alias{addQuotes}
\title{This is a function to add quotation marks around each element of a character string vector}
\usage{
addQuotes(funcIn)
}
\arguments{
\item{funcIn}{a vector of character strings which you want padded by quotation marks}
}
\value{
character vector of length equal to the input
}
\description{
This is a function to add quotation marks around each element of a character string vector
}
