% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{adjustBirths}
\alias{adjustBirths}
\title{This function ensures that a vector of values will sum to a given number.  It's implemented in certain growth forms
(curently: \strong{constant} and \strong{logistic})}
\usage{
adjustBirths(func_adjVector, func_sumTotal,
  func_roundValues = func_roundValues)
}
\arguments{
\item{func_adjVector}{Vector of values which must sum to the func_sumTotal.}

\item{func_sumTotal}{A single integer value which is to be the target summed value.}

\item{func_roundValues}{Logical toggle to control in values must be rounded to integers.}
}
\value{
A vector of values adjusted to sum to a single value.  These may have been forced to be rounded or could still
contain decimals.
}
\description{
This function ensures that a vector of values will sum to a given number.  It's implemented in certain growth forms
(curently: \strong{constant} and \strong{logistic})
}
