% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{buildPedigree}
\alias{buildPedigree}
\title{This is a convenience script to build an named list of empty lists, where the names are based
on the genotype IDs being passed.}
\usage{
buildPedigree(func_focalID)
}
\arguments{
\item{func_focalID}{This should be any vector, that can be interpreted as character, and faithfully represent
the genotype IDs of interest for your pedigree.}
}
\value{
a named list of empty lists.
}
\description{
This is a convenience script to build an named list of empty lists, where the names are based
on the genotype IDs being passed.
}
\examples{
# this creates a named list, this trivial function exists for future flexibility and method design.
buildPedigree(c(1,"zebra","walrus",4))

}
