% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{create_genotypeFrame}
\alias{create_genotypeFrame}
\title{This is a convenience function to ensure that we have a standard shaped data.frame.  It is used to initiate
a new table for the fitness landscape.}
\usage{
create_genotypeFrame(tmpID, tmpStrings, tmpFitnesses)
}
\arguments{
\item{tmpID}{A numeric vector of the unqiue identifiers for genotypes}

\item{tmpStrings}{A vector of the character strings that represent the binary string of genotypes}

\item{tmpFitnesses}{A vector of the numeric fitness values to be input}
}
\value{
A 4 column data frame with column names of genotypeID, binaryString, fitness, isExplored
}
\description{
This is a convenience function to ensure that we have a standard shaped data.frame.  It is used to initiate
a new table for the fitness landscape.
}
\examples{
# This is just a convenience function for outputting vectors in a data.frame with
# standard named columns.
create_genotypeFrame(c(1,10,50),c("1","1_7","6_12"),c(1,0.25,1.57))
}
