% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{expGrowth}
\alias{expGrowth}
\title{This function uses the exponential growth model and can either calculated the expected growth for a single time step
OR it can work backwards to calculated what was the expected starting population size prior to a step of exponential growth.}
\usage{
expGrowth(func_rate, func_step, func_startPop = NULL,
  func_endPop = NULL, func_basalExponent = exp(1))
}
\arguments{
\item{func_rate}{This is the number of offpsring expected to be produced by an individual.  When calculating the expected
population size after a time step, we force this rate to be no less than 1 since this function has meaning only in the birth
function and so we do not want to calculate negative births (which would mean deaths).}

\item{func_step}{This is a proportional scalar that will control what proportion of a standard
"generation" is simulated for each step within a SHAPE run.  NOTE: This parameter is not perfectly
validated to run as may be expected with all models.  For now, it should be left as a value of "1",
but exists for future implementation and testing.}

\item{func_startPop}{This is the initial population size(s) for which you want to calculate a final size.  Leave NULL
if trying to calculated the expected initial size from a final population.}

\item{func_endPop}{This is the final population size(s) for which you want to calculate a initial size.  Leave NULL
if trying to calculated the expected final size from an initial population.}

\item{func_basalExponent}{This controls the basal exponent of the reverse calculation for exponential growth.  It defaults
to the natural exponent and sohuld not be changed from that value.  Modify at your own risk.}
}
\value{
numeric value
}
\description{
This function uses the exponential growth model and can either calculated the expected growth for a single time step
OR it can work backwards to calculated what was the expected starting population size prior to a step of exponential growth.
}
