% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{extract_popDemographics}
\alias{extract_popDemographics}
\title{This is a function that steps forward through time steps of a SHAPE run and extracts population demographic
information.  This includes Fitness, Number of Lineages, and Transitions between dominant genotypes.
Most important it will also return the information related to which lineages will eventually establish in the population,
a piece of information that will be critical for downstream lineage specific information extraction.}
\usage{
extract_popDemographics(func_stepsCon, func_estValue, func_landscapeCon,
  func_hoodCon, func_size_timeStep)
}
\arguments{
\item{func_stepsCon}{This is the filepath to an SQLite database storing information for the stepwise changes of a SHAPE run.}

\item{func_estValue}{This value is used to define the threshold size required for a population before it is considered established.}

\item{func_landscapeCon}{This is the filepath to an SQLite database storing information for the complete explored and neighbouring fitness landscape of a SHAPE run.}

\item{func_hoodCon}{This is the filepath to an SQLite database storing information for high priority mutational neighbourhood information
(which is simply a subset of the full mutational landscape).}

\item{func_size_timeStep}{This is the proportion of a standard biological generation which is to be simulated in a single time step.}
}
\value{
This return a list object that contains various pieces of usefull summary demographic information.
}
\description{
This is a function that steps forward through time steps of a SHAPE run and extracts population demographic
information.  This includes Fitness, Number of Lineages, and Transitions between dominant genotypes.
Most important it will also return the information related to which lineages will eventually establish in the population,
a piece of information that will be critical for downstream lineage specific information extraction.
}
\section{Note}{

There is no example as this cannot work outisde of a runSHAPE call, it requires data produced by the simulation experiment.
}

