% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{lossSampling}
\alias{lossSampling}
\title{This function actually calculates the stochastic loss to populations.}
\usage{
lossSampling(func_inPopulation, func_dilutionFactor)
}
\arguments{
\item{func_inPopulation}{This is a vector of the number of individuals in the populations within the community.}

\item{func_dilutionFactor}{This is expected proportion of the current population sizes that should remain.}
}
\value{
A vector of the resultant population sizes remaining.
}
\description{
This function actually calculates the stochastic loss to populations.
}
\examples{
# A vector of population sizes is randomly sampled to be around the product of size and factor
replicate(5,lossSampling(c(1e4,2e4,3e4),0.01))

}
