% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{name_batchSubmit}
\alias{name_batchSubmit}
\title{This is a function to programatically create R batch submission script names}
\usage{
name_batchSubmit(inVar)
}
\arguments{
\item{inVar}{This is the vector of character string(s) to be used for naming}
}
\value{
A vector of character string of length equal to input.
}
\description{
This is a function to programatically create R batch submission script names
}
