% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{nameEnviron}
\alias{nameEnviron}
\title{This quick little function is a means for me to create the strings of
environments and subsequently extract information back out.}
\usage{
nameEnviron(func_Index, funcSplit = FALSE,
  funcBase = getOption("shape_envString"))
}
\arguments{
\item{func_Index}{This is the vector of numeric, or otherwise unique ID values for the
environments to be created.  Or if funcSplit == TRUE, then these are the names to be split.}

\item{funcSplit}{A logical toggle of whether you are building or splitting the name}

\item{funcBase}{This is the character string used as a prefix to identify environment objects}
}
\value{
A vector of character string of length equal to input.
}
\description{
This quick little function is a means for me to create the strings of
environments and subsequently extract information back out.
}
\examples{
# Returns a standard named string
test_envNames <- nameEnviron(1:10)
nameEnviron(test_envNames, funcSplit = TRUE)
}
