% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{nameObject}
\alias{nameObject}
\title{This quick little function is a means for me to create the strings of
environments and subsequently extract information back out.}
\usage{
nameObject(func_inString, func_inPrefix, func_splitStr = FALSE)
}
\arguments{
\item{func_inString}{This is the vector of numeric, or otherwise unique ID values for the
environments to be created.  Or if funcSplit == TRUE, then these are the names to be split.}

\item{func_inPrefix}{This is the character string used as a prefix to identify environment objects}

\item{func_splitStr}{A logical toggle of whether you are building or splitting the name}
}
\value{
A vector of character string of length equal to input.
}
\description{
This quick little function is a means for me to create the strings of
environments and subsequently extract information back out.
}
\examples{
# Returns a standard named string
test_objectNames <- nameObject(1:10, "testObject")
nameObject(test_objectNames, "testObject", func_splitStr = TRUE)
}
