% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{retrieve_binaryString}
\alias{retrieve_binaryString}
\title{This is a function to search our mutational database and then find the binary string of the genotypeID passed.
This function is more efficient when the number of mutations for each genotypeID be passed as this helps reduce
the tables of the mutational space that are searched.  This matters when large genotypes are simulated.}
\usage{
retrieve_binaryString(func_genotypeID, func_numMuts = NULL,
  func_subNaming, func_landscapeCon)
}
\arguments{
\item{func_genotypeID}{This is a vector of the unique genotype ID for each tracked population in the community.}

\item{func_numMuts}{This is a vector of the number of mutations held within each tracked genotype.}

\item{func_subNaming}{This is a logical which controls if the tables which report on all genotypes with X mutations should be
forced into a single table or it SHAPE is allowed to split them into multiple tables.}

\item{func_landscapeCon}{This is the filepath to an SQLite database storing information for the complete explored and neighbouring fitness landscape of a SHAPE run.}
}
\value{
This returns a vector of character strings that represent the binary strings of the genotypes
}
\description{
This is a function to search our mutational database and then find the binary string of the genotypeID passed.
This function is more efficient when the number of mutations for each genotypeID be passed as this helps reduce
the tables of the mutational space that are searched.  This matters when large genotypes are simulated.
}
\section{Note}{

There is no example as this cannot work outisde of a runSHAPE call, it requires data produced by the simulation experiment.
}

