% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{set_const_NK_interactionsMat}
\alias{set_const_NK_interactionsMat}
\title{This is a function to just return a matrix that defines the sitewise dependencies
for an NK fitness landscape.  If K == 0 or, this is not an NK simulation, it return NULL}
\usage{

  set_const_NK_interactionsMat(func_simModel = getOption("shape_simModel"),
  func_genomeLength = getOption("shape_genomeLength"),
  func_numInteractions = getOption("shape_const_numInteractions"))
}
\arguments{
\item{func_simModel}{This is the fitness landscape model being simulated}

\item{func_genomeLength}{This is the number of sites in the genome being simulated}

\item{func_numInteractions}{An integer value defining the number of sites
that interact with each other site}
}
\value{
Either NULL, or a matrix with K + 1 columns, detailing the sites
interacting with a focal site - identified by the row number and the
cell values of the columns.
}
\description{
This is a function to just return a matrix that defines the sitewise dependencies
for an NK fitness landscape.  If K == 0 or, this is not an NK simulation, it return NULL
}
\section{Note}{

There is no example as this cannot work outisde of a runSHAPE call, it requires data produced by the simulation experiment.
}

