% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{set_const_RMF_globalOptima}
\alias{set_const_RMF_globalOptima}
\title{This function samples the space of all possible genotypes and then defines
one that will be considered as the independent fitness contribution global optima.}
\usage{
set_const_RMF_globalOptima(func_simModel = getOption("shape_simModel"),
  func_genomeLength = getOption("shape_genomeLength"),
  func_initDistance = getOption("shape_const_RMF_initiDistance"),
  func_sepString = getOption("shape_sepString"))
}
\arguments{
\item{func_simModel}{This is the fitness landscape model being simulated}

\item{func_genomeLength}{The number of sites in the genome being simulated}

\item{func_initDistance}{This is the number of mutations found in the global optimal genotype}

\item{func_sepString}{This is the string collapse separator used in the run}
}
\value{
A character string of genome positions at which there ought to be mutations to be optimal
}
\description{
This function samples the space of all possible genotypes and then defines
one that will be considered as the independent fitness contribution global optima.
}
\section{Note}{

There is no example as this cannot work outisde of a runSHAPE call, it requires data produced by the simulation experiment.
}

