% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{set_DepbySite_ancestFitness}
\alias{set_DepbySite_ancestFitness}
\title{This is a convenience function for setting the dependent fitness values of sites in
an NK fitness landscape model.  This allows the dependent fitness of sites to be calculated
once and then referenced as mutations occur.  It makes exploring this style of fitness landscape
a bit more computationally friendly - as it generally isn't.}
\usage{
set_DepbySite_ancestFitness(func_simModel = getOption("shape_simModel"),
  func_const_siteBystate_fitnessMat = getOption("shape_const_siteBystate_fitnessMat"),
  func_const_NK_interactionMat = getOption("shape_const_NK_interactionMat"))
}
\arguments{
\item{func_simModel}{This is the fitness landscape model being simulated}

\item{func_const_siteBystate_fitnessMat}{This is the sitewise independent fitness contributions in the fitness landscape}

\item{func_const_NK_interactionMat}{This defines the sitewise dependencies based on the K interactions.}
}
\value{
Either the dependent sitewise fitness contributions in an NK fitness landscape, or NA.
}
\description{
This is a convenience function for setting the dependent fitness values of sites in
an NK fitness landscape model.  This allows the dependent fitness of sites to be calculated
once and then referenced as mutations occur.  It makes exploring this style of fitness landscape
a bit more computationally friendly - as it generally isn't.
}
\section{Note}{

There is no example as this cannot work outisde of a runSHAPE call, it requires data produced by the simulation experiment.
}

