% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{shapeCombinations}
\alias{shapeCombinations}
\title{This is a function to take the input parameters and build the parameter combinations}
\usage{
shapeCombinations(func_inLines, func_comboRef, func_indepRef, func_condRef)
}
\arguments{
\item{func_inLines}{These are the template lines of text to be updated.}

\item{func_comboRef}{This is the reference identifiers for grouped as pairwise parameter combinations}

\item{func_indepRef}{This is the reference identifiers for independent parameter values not to be done pairwise}

\item{func_condRef}{This is the reference indetifiers for grouped parameter combinations which are conditional on others.}
}
\value{
A table of parameter combinations which represents the combination of experimental parameters for a SHAPE experiment.
}
\description{
This is a function to take the input parameters and build the parameter combinations
}
\section{Note}{

There is no example as this cannot work outisde of a runSHAPE call, it requires data produced by the simulation experiment.
}

