% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{stopError}
\alias{stopError}
\title{This is a convenience wrapper for sending an error and ending the SHAPE run as well as the R environment.
It will print a message and then traceback() report before pausing and quiting the R session.  This exists
to help debugging when SHAPE is run in batch-mode.}
\usage{
stopError(func_message)
}
\arguments{
\item{func_message}{The message to be sent to screen prior to ending the R session.}
}
\description{
This is a convenience wrapper for sending an error and ending the SHAPE run as well as the R environment.
It will print a message and then traceback() report before pausing and quiting the R session.  This exists
to help debugging when SHAPE is run in batch-mode.
}
\section{Note}{

There is no example as this functions role is to print a message
and then quit the R run.
}

