% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{summarise_experimentFiles}
\alias{summarise_experimentFiles}
\title{This function will find all initially processed output files from individual replicates and return summary information.
That information is saved to an RData file which will contain 3 objects: all_proccessedFiles, all_jobInfo, all_dividedFiles}
\usage{

  summarise_experimentFiles(func_experimentDir = getOption("shape_workDir"),
  func_saveFile = getOption("shape_procExp_filenames")["fileList"],
  func_search_filePattern = getOption("shape_processedData_filePattern"),
  func_sepString = getOption("shape_sepString"))
}
\arguments{
\item{func_experimentDir}{This is the filepath to the root directoy under which all your experimental files can
be found.}

\item{func_saveFile}{This is the filepath and filename (ending in .RData please) to which the results of this
step will be saved.}

\item{func_search_filePattern}{This is a string which can be used to search and find the files which relate to
the processed output of individual replicates rSHAPE runs.}

\item{func_sepString}{This is the character string which was used for commonly collapsing elements in the rSHAPE run.}
}
\description{
This function will find all initially processed output files from individual replicates and return summary information.
That information is saved to an RData file which will contain 3 objects: all_proccessedFiles, all_jobInfo, all_dividedFiles
}
\section{Note}{

There is no example as this cannot work without a complete rSHAPE experiment to be analysed.
}

