% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{summarise_experimentParameters}
\alias{summarise_experimentParameters}
\title{This function will use output from summarise_experimentFiles to locate all parameter files and then
report on all those parameters for the jobs that were run.  This will save an RData file which will
contain one object: all_parmInfo}
\usage{
summarise_experimentParameters(func_workEnvir = new.env(),
  func_saveFile = getOption("shape_procExp_filenames")["parameters"],
  func_experimentDir = getOption("shape_workDir"),
  func_refFile = getOption("shape_procExp_filenames")["fileList"])
}
\arguments{
\item{func_workEnvir}{This is an environment used to load files with the load function.  It's used to encapsulate
the loaded information to a controlled space.}

\item{func_saveFile}{This is the filepath and filename (ending in .RData please) to which the results of this
step will be saved.}

\item{func_experimentDir}{This is the filepath to the root directoy under which all your experimental files can
be found.}

\item{func_refFile}{This is the filepath to the reference file that contains information regarding all the
processed files for the rSHAPE experiment.}
}
\description{
This function will use output from summarise_experimentFiles to locate all parameter files and then
report on all those parameters for the jobs that were run.  This will save an RData file which will
contain one object: all_parmInfo
}
\section{Note}{

There is no example as this cannot work without a complete rSHAPE experiment to be analysed.
}

