% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{trimQuotes}
\alias{trimQuotes}
\title{This is a function to trim a string by removing the first and last character, it's used to trim quotation marks
used in the parameter input}
\usage{
trimQuotes(funcIn)
}
\arguments{
\item{funcIn}{a vector of character strings which you want trimmed}
}
\value{
character vector of length equal to the input
}
\description{
This is a function to trim a string by removing the first and last character, it's used to trim quotation marks
used in the parameter input
}
\examples{
# It removes leading and trailing string positions, use when quotations are known to exist.
trimQuotes(c('"someWords"','otherwords"',"is_changed"))

}
