% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{updateLines}
\alias{updateLines}
\title{This is a function which is used to update lines that are searched and replace in a manner conditional to this script's circumstances
The input lines can be a vector of any length, and the search patterns can be a list of any length where each list vector is used together.
The values should be a list of information used as replacement info.}
\usage{
updateLines(func_inLines, func_searchPattern, func_values)
}
\arguments{
\item{func_inLines}{These are the lines that are to be updated before output}

\item{func_searchPattern}{These are the string-s- to be searched for replacement}

\item{func_values}{These are the values that are to replace the searched strings.}
}
\value{
A vector of character strings that has now been updated.
}
\description{
This is a function which is used to update lines that are searched and replace in a manner conditional to this script's circumstances
The input lines can be a vector of any length, and the search patterns can be a list of any length where each list vector is used together.
The values should be a list of information used as replacement info.
}
\section{Note}{

There is no example as this cannot work outisde of a runSHAPE call, it requires data produced by the simulation experiment.
}

