% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSHAPE.R
\name{writeParameters}
\alias{writeParameters}
\title{This is a file for updating the post analysis plotting script and creating an updated copy in the experiment's folder}
\usage{
writeParameters(func_infile, func_inParms, func_inCombos, func_outDir,
  func_bodyScript, func_ExternalStopper,
  func_sepString = getOption("shape_sepString"))
}
\arguments{
\item{func_infile}{This is the filepath location for the template script to be writte in.}

\item{func_inParms}{These are the parameters to be updated in the plotting file}

\item{func_inCombos}{This is the combination of parameters to be written}

\item{func_outDir}{This is the director filepath to which output should be written.}

\item{func_bodyScript}{This is a run body of SHAPE script to be read in as template}

\item{func_ExternalStopper}{This is a file placed externally used as a logical flag that SHAPE should
stop trying to seed new replicates to be run.}

\item{func_sepString}{This is the common string for collapsing information.}
}
\value{
A character string indicating that the plotting file-s- have been written
}
\description{
This is a file for updating the post analysis plotting script and creating an updated copy in the experiment's folder
}
\section{Note}{

There is no example as this cannot work outisde of a runSHAPE call, it requires data produced by the simulation experiment.
}

